animal routes

+++ api/src/modules/animal/useCases/animal/deleteAnimal/DeleteAnimalUseCase.ts
@@ -1,8 +1,11 @@
 import { IAnimalRepository } from "@modules/animal/repositories/IAnimalRepository";
 import { inject, injectable } from "tsyringe";
 
+import { AppError } from "@shared/errors/AppError";
+
 interface IRequest {
     cod_animal: number;
+    cod_usuario: number;
 }
 
 @injectable()
@@ -12,7 +15,20 @@ class DeleteAnimalUseCase {
         private animalRepository: IAnimalRepository
     ) {}
 
-    async execute({ cod_animal }: IRequest): Promise<void> {
+    async execute({ cod_animal, cod_usuario }: IRequest): Promise<void> {
+        const animalExists = await this.animalRepository.findById(cod_animal);
+
+        if (!animalExists) {
+            throw new AppError(
+                "O código de animal informado não foi localizado na lista."
+            );
+        }
+        if (animalExists.cod_usuario !== cod_usuario) {
+            throw new AppError(
+                "Não é possivel apagar o animal cadastrado por outro usuário.",
+                401
+            );
+        }
         await this.animalRepository.delete(cod_animal);
     }
 }
